docs and move event bus tests
