+++ test/integration/kafka/docker-compose.yaml
